Added AnimeInfoExtractor.py testing script
anime_extractor_test allows for testing a spreadsheet of filenames to see what information Trackma's AnimeInfoExtractor will pull from each one, and how it ends up with the information it does by showing the output of each function it runs. This information is written to the 'summary.xlsx' and 'output.xlsx' spreadsheets, respectively.